HACK: usb: hub: disable autosuspend for Genesys Logic Hubs
Disable autosuspend in Genesys Logic hubs to allow USB devices on the
Odroid C2 board to be used. The alternative to this patch is setting
usbcore.autosuspend=-1 in boot params.

This patch only impacts GXBB devices as GXL/GXM onwards use the newer
dwc3 core which does not have the problem.

Signed-off-by: Christian Hewitt <christianshewitt@gmail.com>