johan_defconfig: add custom config
Add a custom trimmed configuration that can be used to test SC8280XP
support including the Lenovo ThinkPad X13s, sc8280xp-crd and
sa8295p-adp.

Make sure the initramfs includes any modules required to boot, for
example:

	nvme phy_qcom_qmp_pcie pcie_qcom

for the X13s, and

	phy_qcom_qmp_ufs ufs_qcom

for the CRD with rootfs on UFS.

For keyboard input and (more than 30 seconds of) display in initramfs,
make sure to also include:

	i2c_hid_of i2c_qcom_geni

for keyboard, and

	leds_qcom_lpg pwm_bl
	qrtr pmic_glink_altmode gpio_sbu_mux phy_qcom_qmp_combo
	panel_edp msm phy_qcom_edp

for the display (preferably in the above order to avoid an msm drm
probe-deferral issue).

Not-signed-off-by: Johan Hovold <johan+linaro@kernel.org>