arm64: dts: qcom: hack in sa8540p support
The SA8540P has 4 clocks listed for the UFS PHY in the downstream
DTS, whereas the sa8295p and sc8280xp has two listed. Modify the
existing DTS files and hack in the expected number of clocks.

With the clock changes, the sa8540p is now able to mount the UFS and
boot directly into a CS9 userspace.

    CentOS Stream 9
    Kernel 5.17.0-rc7-next-20220310-11984-g14a06258b39b-dirty on an aarch64

    localhost login:

The regulator names matched, however the vccq-max-microamp is lower
for this platform in the downstream DTS, so lower it here to be safe.

Signed-off-by: Brian Masney <bmasney@redhat.com>