phy: qcom-qmp-pcie: add support for pipediv2 clock
Some QMP PHYs have a second fixed-divider pipe clock that needs to be
enabled along with the pipe clock.

Add support for an optional "pipediv2" clock.

Note that there's no point in adding a "lane" suffix to the clock name
as there will only ever be one pipe clock per PHY (the same is true for
the "pipe0" clock which could be renamed "pipe").

Signed-off-by: Johan Hovold <johan+linaro@kernel.org>