wip! firmware: Add support for Qualcomm Trusted Execution Environment SCM calls
Add support for SCM calls to Secure OS and the Trusted Execution
Environment (TEE/TrEE) residing in the TrustZone (TZ). This allows
communication with Secure/TZ applications, for example 'uefisecapp'
managing access to UEFI variables.

While this patch introduces only a very basic interface without the more
advanced features (such as reentrant and blocking SCM calls and
listeners/callbacks), this is enough to talk to the aforemetioned
'uefisecapp'.

Signed-off-by: Maximilian Luz <luzmaximilian@gmail.com>