arm64: dts: qcom: sc8280xp: Switch to GICv3-ITS for MSIs
Instead of using the DWC internal MSI controller, let's use the GICv3-ITS for
handling the MSIs from PCIe controllers. This has the advantage of assigning
SMP affinity to each IRQs.

TODO: The node name gic-its has been changed to msi-controller. But this is
still not documented in devicetree spec.

Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>