hack: arm64: dts: qcom: sc8280xp-x13s: enable both touchpads
Enable both touchpad nodes in the devictree and let the HID driver
determine which one is actually populated.

Note that the pin configuration must currently be moved to the parent
i2c-bus node even though only one of these nodes will ever be
successfully probed on a specific device.

This is needed until the boot firmware can enable only the node for the
populated touchpad.

Not-signed-off-by: Johan Hovold <johan+linaro@kernel.org>