btrfs: send: add support for fs-verity
Preserve the fs-verity status of a btrfs file across send/recv.

There is no facility for installing the Merkle tree contents directly on
the receiving filesystem, so we package up the parameters used to enable
verity found in the verity descriptor. This gives the receive side
enough information to properly enable verity again. Note that this means
that receive will have to re-compute the whole Merkle tree, similar to
how compression worked before encoded_write.

Since the file becomes read-only after verity is enabled, it is
important that verity is added to the send stream after any file writes.
Therefore, when we process a verity item, merely note that it happened,
then actually create the command in the send stream during
'finish_inode_if_needed'.

This also creates V3 of the send stream format, without any format
changes besides adding the new commands and attributes.

Signed-off-by: Boris Burkov <boris@bur.io>

--
Changes in v4:
- Use btrfs_get_verity_descriptor instead of verity ops get descriptor.
  Move that definition to ctree.h for conditional building. This cleaned
  up most of the conditional building issues, in my opinion.
- Rename process_new_verity to process_verity.
- Use le-to-cpu conversion for all fsverity descriptor fields.
- Don't check NULL for kvfree of the send descriptor.
Changes in v3:
- Fixed build failure when CONFIG_FS_VERITY was not set. This required a
  kludge to avoid a build warning as well.
Changes in v2:
- Allocate 16K with kvmalloc and keep it around till the end of send
instead of re-allocating on each file with fs-verity.
- Use unsigned literal for bitshift.

Signed-off-by: David Sterba <dsterba@suse.com>