usb: dwc3: keep PHYs disabled during suspend
Commit 649f5c842ba3 ("usb: dwc3: core: Host wake up support from system
suspend") started leaving the PHYs enabled during suspend for
wakeup-capable controllers even though it turns out this had nothing to
do with wakeup.

Rather, the wakeup capability flag was (ab-)used as a proxy to configure
the suspend behaviour in an attempt to reduce power leakage on some
platforms.

Stop abusing the wakeup configuration and restore the 5.19 behaviour of
keeping the PHYs powered off during suspend. If needed, a dedicated
mechanism for configuring the PHY power state during suspend can be
added later.

Fixes: 649f5c842ba3 ("usb: dwc3: core: Host wake up support from system suspend")
Link: https://lore.kernel.org/r/Yuv7AM/5jtO/pgcm@google.com
Signed-off-by: Johan Hovold <johan+linaro@kernel.org>