ASoC: codecs: lpass: do not handle frame sync clock in runtime pm
Do not handle frame sync clock in runtime pm as this will put the
devices on the bus out of sync, resulting in a re-enumeration.

Lossing devices from bus in the middle of runtime pm resume
can have very bad side effects as we will be handling regmap cache
with old device id and this could underflow read fifo and overflow
write fifos.

Signed-off-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>